Rebalanced block toughness, tool durability, and tool multipliers
-also updated the DamageAspect to use ints, intead of shorts, as we now want to use the extra precision the int gives us
-primarily, this has the effect of dramatically increasing block hardness and a corresponding increase in tool durability (tools from previous versions will see a durability drop)
-the general intension is to encourage the use of tools, and higher-tier tools, as there was little need to use them with previous hardness

+++ october-common/src/main/java/com/jeffdisher/october/mutations/MutationBlockIncrementalBreak.java
@@ -61,11 +61,11 @@ public class MutationBlockIncrementalBreak implements IMutationBlock
 		
 		// We want to see if this is a kind of block which can be broken.
 		Block block = newBlock.getBlock();
-		short toughness = env.damage.getToughness(block);
+		int toughness = env.damage.getToughness(block);
 		if (DamageAspect.UNBREAKABLE != toughness)
 		{
 			// Apply the damage.
-			short damage = (short)(newBlock.getDamage() + _damageToApply);
+			int damage = (newBlock.getDamage() + _damageToApply);
 			
 			// See if this is broken (note that damage could overflow).
 			if ((damage >= toughness) || (damage < 0))
